+++ RELEASE_NOTES.md
@@ -62,6 +62,9 @@ Tests
 * IMPROVED: The `logtalk_tester` shell script now supports running unit tests
 in `optimal` mode, besides the already supported `normal` and `debug` modes.
 
+* IMPROVED: The `logtalk_tester` shell script now prints the total number of
+tests skipped, passed, and failed.
+
 * ADDED: Unit tests for the `catch/3`, `forall/2`, and `(\+)/1` built-in
 methods.